refactor(storefront/cart): move useEnrichedCart hook to shared/lib
